@@ -19,33 +19,33 @@
     "generate-pwa-assets": "pwa-assets-generator --preset minimal-2023 static/images/icons/favicon.svg"
   },
   "devDependencies": {
-    "@playwright/test": "^1.47.2",
+    "@playwright/test": "^1.48.1",
     "@skeletonlabs/skeleton": "^2.10.2",
     "@skeletonlabs/tw-plugin": "~0.4.0",
-    "@sveltejs/adapter-node": "^5.2.4",
-    "@sveltejs/kit": "^2.5.28",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/adapter-node": "^5.2.8",
+    "@sveltejs/kit": "^2.7.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/forms": "~0.5.9",
     "@tailwindcss/typography": "~0.5.15",
     "@types/canvas-confetti": "^1.6.4",
-    "@types/node": "^22.5.5",
+    "@types/node": "^22.7.7",
     "@types/uuid": "^10.0.0",
     "@vite-pwa/assets-generator": "~0.2.6",
-    "@vite-pwa/sveltekit": "^0.6.5",
+    "@vite-pwa/sveltekit": "^0.6.6",
     "autoprefixer": "^10.4.20",
     "editorconfig": "^2.0.0",
     "postcss": "^8.4.47",
     "postcss-load-config": "^6.0.1",
-    "svelte": "^4.2.19",
-    "svelte-check": "^4.0.2",
-    "svelte-dnd-action": "~0.9.50",
-    "tailwindcss": "^3.4.12",
-    "tslib": "^2.7.0",
-    "typescript": "^5.6.2",
-    "vite": "^5.4.7",
+    "svelte": "^5.0.2",
+    "svelte-check": "^4.0.5",
+    "svelte-dnd-action": "~0.9.52",
+    "tailwindcss": "^3.4.14",
+    "tslib": "^2.8.0",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.9",
     "vite-plugin-pwa": "~0.20.5",
     "vite-plugin-tailwind-purgecss": "~0.3.3",
-    "vitest": "^2.1.1",
+    "vitest": "^2.1.3",
     "workbox-window": "^7.1.0"
   },
   "dependencies": {
@@ -53,8 +53,8 @@
     "@napi-rs/canvas": "~0.1.56",
     "browser-fs-access": "~0.35.0",
     "canvas-confetti": "^1.9.3",
-    "firebase": "^10.13.2",
-    "firebase-admin": "^12.5.0",
+    "firebase": "^10.14.1",
+    "firebase-admin": "^12.6.0",
     "uuid": "^10.0.0",
     "validate-color": "^2.2.4",
     "zod": "^3.23.8"
