@@ -11,9 +11,9 @@
   const modalStore = getModalStore()
   const toastStore = getToastStore()
 
-  let title = $wheelStore.config.title
-  let loading = false
-  let saveMode: 'overwrite' | 'new' = 'new'
+  let title = $state($wheelStore.config.title)
+  let loading = $state(false)
+  let saveMode: 'overwrite' | 'new' = $state('new')
 
   onMount(async () => {
     const user = getCurrentUser()
@@ -43,7 +43,8 @@
     }
   })
 
-  const save = async () => {
+  const save = async (e: Event) => {
+    e.preventDefault()
     if (loading) return
     loading = true
     try {
@@ -112,7 +113,7 @@
     </header>
 
     <form
-      on:submit|preventDefault={save}
+      onsubmit={save}
       class="flex flex-col gap-4"
     >
       {#if $wheelStore.path}
@@ -159,7 +160,7 @@
         <button
           type="button"
           class="btn variant-soft"
-          on:click={modalStore.close}
+          onclick={modalStore.close}
         >
           Cancel
         </button>
